DOS/soa/service/requirements.txt
flask-restplus
gunicorn
psycopg2
Werkzeug==0.16.1
PyJWT
cryptography
